DDI_backend_final/requirements.txt
# Core Django packages
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dj-rest-auth==7.0.1
django-allauth==65.10.0
django-cors-headers==4.9.0
django-filter==25.1
drf-spectacular==0.28.0

# Database
dj-database-url==2.1.0
psycopg2-binary==2.9.10

# Email
sendgrid-django==4.2.0

# Environment
python-dotenv==1.1.1

# Production server
gunicorn==21.2.0
whitenoise==6.11.0

# Image processing
Pillow==11.3.0

# WebAuthn
webauthn==2.7.0

# 2FA and QR codes
pyotp==2.9.0
qrcode==7.4.2

# HTTP client (for HF API calls)
httpx==0.28.1

# Gradio client (for Hugging Face API)
gradio-client==1.13.0

# Additional packages for HF API
huggingface-hub==0.35.0
fsspec==2025.9.0
tqdm==4.67.1

# Utility packages
requests==2.32.5
PyYAML==6.0.2
